requirements.txt
asgiref==3.7.2
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
Pillow==10.0.1
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.7.0
